% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_distance.R
\name{Calculate_distance}
\alias{Calculate_distance}
\title{Calculate gene pairwise distance}
\usage{
Calculate_distance(dat, method)
}
\arguments{
\item{dat}{matrix; expression matrix}

\item{method}{string; metric for measuring gene-gene distance (option: pearson, euclidean, KL, jaccard, spearman)}
}
\value{
A distance matrix
}
\description{
Calculate gene pairwise distance
}
\examples{


}
