% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_gene_module.R
\name{ClusterGenes}
\alias{ClusterGenes}
\title{Cluster Genes}
\usage{
ClusterGenes(
  dist,
  clustering_method = "average",
  min_gene = 10,
  deepSplit = 2,
  return_tree = FALSE,
  filtered = TRUE,
  accu = 0.75
)
}
\arguments{
\item{dist}{dist; the gene-gene distance matrix.}

\item{clustering_method}{character; the method for clustering the genes. Options are "average", "ward.D", "ward.D2", "single", "complete", "mcquitty", "median", "centroid", "dbscan", or "hdbscan". Default is "average".}

\item{min_gene}{integer; the minimum number of genes each group should contain. Default is 10.}

\item{deepSplit}{integer; parameters for \code{cutreeDynamic}. Default is 2.}

\item{return_tree}{logical; if TRUE, returns a gene partition tree; otherwise returns only the gene partition. Default is FALSE.}

\item{filtered}{logical; if TRUE, filters out some genes for each partition based on the SML method (Parisi et al., 2014). Default is TRUE.}

\item{accu}{numeric; the threshold for filtering out genes. Default is 0.75.}
}
\value{
If \code{return_tree} is TRUE, returns a list containing the gene partition and the gene partition tree. Otherwise, returns the gene partition.
}
\description{
This function partitions genes based on their pairwise distances using various clustering methods.
}
\details{
This function performs hierarchical clustering using the specified method and partitions the genes. It also provides options for other clustering methods like dbscan and hdbscan, and to filter out noisy genes.
}
\examples{
gene_dist <- dist(matrix(runif(100), nrow=10))
gene_partition <- ClusterGenes(gene_dist, clustering_method="average")


}
\references{
Parisi, F., Strino, F., Nadler, B., & Kluger, Y. (2014). Ranking and combining multiple predictors without labeled data. Proceedings of the National Academy of Sciences, 111(4), 1253-1258. doi:10.1073/pnas.1219097111
}
