% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_cell_graph.R
\name{ConstructDiffusionOperators}
\alias{ConstructDiffusionOperators}
\title{Construct Diffusion Operators}
\usage{
ConstructDiffusionOperators(W, max_time)
}
\arguments{
\item{W}{matrix; symmetric affinity matrix of the graph.}

\item{max_time}{integer; the maximum diffusion time. The actual maximum diffusion time may be shorter if all nodes converge beforehand.}
}
\value{
A list of sparse matrices representing the diffusion operators at different times.
Each element in the list corresponds to P^t for t = 0, 2, 4, 8, ..., max_time.
The list is named by the corresponding diffusion time.
}
\description{
Constructs a list of diffusion operators for a given symmetry affinity matrix of a graph.
}
\keyword{internal}
