% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_cell_graph.R
\name{ConstructGaussianGraph}
\alias{ConstructGaussianGraph}
\title{Construct a Symmetric Gaussian Graph}
\usage{
ConstructGaussianGraph(
  knn = 5,
  feature_space,
  alpha = 1,
  coef = 2,
  epsilon = 0.001,
  self_loop = 1
)
}
\arguments{
\item{knn}{integer; the number of nearest neighbors to consider for each node. Default is 5.}

\item{feature_space}{matrix; a cell-by-coordinate matrix (e.g., 20 principal components)}

\item{alpha}{numeric; the exponent in the Gaussian kernel computation. Default is 1.}

\item{coef}{numeric; used to compute the adaptive bandwidth. Default is 2.}

\item{epsilon}{numeric; threshold below which edge weights are set to zero to sparsify the graph. Default is 1e-3.}

\item{self_loop}{numeric; value for the diagonal elements of the affinity matrix to add self-loops. Default is 1.}
}
\value{
A list containing the following elements:
\describe{
\item{graph}{A symmetric matrix representing the graph affinity matrix.}
\item{adj_matrix}{A binary adjacency matrix indicating the presence of edges between nodes.}
}
}
\description{
Constructs a Gaussian kernel graph from a given feature space.
}
\examples{
\dontrun{
feature_space <- matrix(rnorm(100), nrow = 10)
result <- ConstructGaussianGraph(knn = 3, feature_space = feature_space)
graph <- result$graph
adj_matrix <- result$adj_matrix
}

}
\keyword{internal}
