% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_cell_graph.R
\name{ConstructKnnGraph}
\alias{ConstructKnnGraph}
\title{Construct a Symmetric KNN graph}
\usage{
ConstructKnnGraph(knn = 5, feature_space, adjust_disconnection = TRUE, self_loop = 1)
}
\arguments{
\item{knn}{integer; the number of nearest neighbors to consider for constructing the graph. Default is 5.}

\item{feature_space}{matrix; a cell-by-coordinate matrix (e.g., 20 principal components).}

\item{adjust_disconnection}{boolean; TRUE for connecting disconnected components using Minimum Spanning Trees (MST). Default is TRUE.}

\item{self_loop}{integer; weight for self connections (default is 1).}
}
\value{
A list containing the following elements:
\describe{
\item{graph}{matrix; symmetric KNN graph W, computed as \code{pmax(1, (A + A^T) / 2)}.}
\item{adj_matrix}{matrix; adjacency matrix A.}
\item{component}{list; disconnected components in the graph, each component is represented as a subgraph.}
}
}
\description{
Constructs a symmetric KNN graph from a given feature space.
}
\examples{
feature_space <- matrix(rnorm(200), nrow = 10, ncol = 20)
result <- ConstructKnnGraph(knn = 3, feature_space = feature_space)
graph <- result$graph
adj_matrix <- result$adj_matrix

}
