% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_cell_graph.R
\name{ConstructSNNGraph}
\alias{ConstructSNNGraph}
\title{Construct a Symmetric SNN Graph}
\usage{
ConstructSNNGraph(
  knn = 5,
  feature_space,
  self_loop = 1,
  adjust_disconnection = TRUE
)
}
\arguments{
\item{knn}{integer; the number of nearest neighbors to consider for each node. Default is 5.}

\item{feature_space}{matrix; a cell-by-coordinate matrix (e.g., 20 principal components)}

\item{self_loop}{numeric; value for the diagonal elements of the affinity matrix to add self-loops. Default is 1.}
}
\value{
A list containing the following elements:
\describe{
\item{graph}{A symmetric matrix representing the graph affinity matrix.}
\item{adj_matrix}{A binary adjacency matrix indicating the presence of edges between nodes.}
}
}
\description{
Constructs a Shared Nearest Neighbor (SNN) graph from a given feature space using k-nearest neighbors (KNN) and Seurat's SNN computation.
}
\keyword{internal}
