% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CustomFeaturePlot}
\alias{CustomFeaturePlot}
\title{Custom Feature Plot}
\usage{
CustomFeaturePlot(
  coord,
  value,
  reduction = NULL,
  dims = 1:2,
  value_name = NULL,
  title_name = NULL,
  order_point = TRUE
)
}
\arguments{
\item{coord}{A Seurat object or a matrix of coordinates. If a Seurat object is provided, embeddings will be extracted from Seurat.}

\item{value}{A vector of values to plot. If \code{coord} is a Seurat object, \code{value} can be a column name from \code{coord@meta.data}.}

\item{reduction}{Character; the name of the reduction embedding to use (e.g., "umap", "tsne"). Default is NULL.}

\item{dims}{A vector specifying which dimensions to use. Default is c(1,2).}

\item{value_name}{Character; the name of the value to be used in the plot legend. Default is NULL.}

\item{title_name}{Character; the title of the plot. Default is NULL.}

\item{order_point}{Logical; whether to order points by value. Default is TRUE.}
}
\value{
A ggplot2 object representing the custom feature plot.
}
\description{
Creates a custom feature plot for visualizing a feature's values on a reduced dimension embedding (e.g., UMAP or t-SNE).
}
