% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CustomModulePlot}
\alias{CustomModulePlot}
\title{Plot Feature Modules}
\usage{
CustomModulePlot(
  dat,
  coord = NULL,
  feature_partition,
  reduction = NULL,
  assays = "RNA"
)
}
\arguments{
\item{dat}{A Seurat object or a matrix of gene expression data.}

\item{coord}{matrix; 2D coordinates of cells for visualization. Required if \code{dat} is not a Seurat object.}

\item{feature_partition}{factor; a vector indicating the partition of features (genes) into modules.}

\item{reduction}{character; the name of the reduction embedding to use (e.g., "umap", "tsne"). Default is NULL.}

\item{assays}{character; the assay to use from the Seurat object. Default is "RNA".}
}
\value{
A list of ggplot2 objects representing the feature plots for each module.
}
\description{
Creates feature plots for visualizing the average expression of gene modules across cells.
}
