% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lmds.R
\name{fast_calculate_score_profile}
\alias{fast_calculate_score_profile}
\title{Fast Calculate Score Profile}
\usage{
fast_calculate_score_profile(
  W,
  max_time = 2^15,
  init_state,
  P_ls = NULL,
  score_ls = c("score0")
)
}
\arguments{
\item{W}{matrix; the affinity matrix.}

\item{max_time}{integer; the maximum diffusion time. Default is 2^15. The actual maximum diffusion time may be shorter if all nodes converge beforehand.}

\item{init_state}{matrix; the row-wise normalized initial state matrix.}

\item{P_ls}{list; the list of diffusion operators. Default is NULL.}

\item{score_ls}{character vector; the list of scores to compute. See \code{\link{get_score_profile}} for details. Default is \code{c("score0")}.}
}
\value{
A data frame containing the computed scores for different diffusion times.
}
\description{
Calculates the multi-scale KL divergence score profile with dyadic time scales (0, 2, 4, ...).
}
\keyword{internal}
