% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lmds.R
\name{fast_calculate_score_profile_highres}
\alias{fast_calculate_score_profile_highres}
\title{Fast Calculate Score Profile with Fine Time Scale}
\usage{
fast_calculate_score_profile_highres(
  W,
  max_time = 100,
  init_state,
  score_ls = c("score0")
)
}
\arguments{
\item{W}{matrix; the affinity matrix.}

\item{max_time}{integer; the maximum diffusion time. Default is 100. The actual maximum diffusion time may be shorter if all nodes converge beforehand.}

\item{init_state}{matrix; the row-wise normalized initial state matrix.}

\item{score_ls}{character vector; the list of scores to compute. See \code{\link{get_score_profile}} for details. Default is \code{c("score0")}.}
}
\value{
A data frame containing the computed scores for different diffusion times.
}
\description{
Calculates the multi-scale KL divergence score profile with fine time scales (0, 1, 2, 3, ...).
}
\keyword{internal}
