% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lmds.R
\name{fast_get_lmds}
\alias{fast_get_lmds}
\title{Fast Get LMDS}
\usage{
fast_get_lmds(
  W,
  max_time = 2^15,
  init_state,
  P_ls = NULL,
  correction = FALSE,
  largeData = TRUE,
  highres = FALSE
)
}
\arguments{
\item{W}{matrix; the affinity matrix.}

\item{max_time}{integer; the maximum diffusion time. Default is 2^15. The actual maximum diffusion time may be shorter if all nodes converge beforehand.}

\item{init_state}{matrix; the row-wise normalized expression matrix.}

\item{P_ls}{list; the list of diffusion operators. Default is NULL.}

\item{correction}{logical; if TRUE, adjusts the score profile by scaling the score profile by entropy and applying delta correction. Default is FALSE.}

\item{largeData}{logical; if TRUE, uses \code{\link{fast_calculate_score_profile_largeData}}. Default is TRUE.}

\item{highres}{logical; if TRUE, uses \code{\link{fast_calculate_score_profile_highres}}. Default is FALSE.}
}
\value{
A list containing:
\item{score_profile}{data frame; the computed score profile for different diffusion times.}
\item{cumulative_score}{numeric vector; the LMDS for each row in the score profile.}
}
\description{
Fast Get LMDS
}
