% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{fastKLMatrix}
\alias{fastKLMatrix}
\title{Fast KL Divergence Calculation for Matrices}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{A numeric matrix.}
}
\value{
A numeric vector with the KL divergence values.
}
\description{
This function calculates the KL divergence between rows of two matrices.
}
\examples{
x <- matrix(runif(10), 5, 2)
y <- matrix(runif(10), 5, 2)
fastKLMatrix(x, y)
}
\keyword{internal}
