% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{fastKLVector}
\alias{fastKLVector}
\title{Fast KL Divergence Calculation for Matrix and Vector}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{A numeric vector.}
}
\value{
A numeric vector with the KL divergence values.
}
\description{
This function calculates the KL divergence between rows of a matrix and a vector.
}
\examples{
x <- matrix(runif(10), 5, 2)
y <- runif(2)
fastKLVector(x, y)
}
\keyword{internal}
