% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{fastMatMult}
\alias{fastMatMult}
\title{Fast Matrix Multiplication}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}
}
\value{
A numeric matrix which is the product of A and B.
}
\description{
This function multiplies two matrices using Armadillo.
}
\examples{
A <- matrix(1:4, 2, 2)
B <- matrix(5:8, 2, 2)
fastMatMult(A, B)
}
\keyword{internal}
