% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturePlot_diffusion.R
\name{FeaturePlot_diffusion}
\alias{FeaturePlot_diffusion}
\title{Visualize diffusion}
\usage{
FeaturePlot_diffusion(
  coord,
  init_state,
  P_ls = NULL,
  W = NULL,
  check_time = NULL,
  gene_name = NULL,
  gene_color = "blue"
)
}
\arguments{
\item{coord}{matrix; 2D coordinate of cells for visualization}

\item{init_state}{matrix; Row-wise normalized expression matrix}

\item{P_ls}{matrix; Diffused transition matrix}

\item{W}{matrix; Symmetry affinity matrix of kNN graph}

\item{check_time}{vector; Select several time point for visualization, e.g. c(2,16,128)}

\item{gene_name}{string; Select one gene for visualization}

\item{gene_color}{string; The color of expression level for visualization}
}
\value{
A ggplot object
}
\description{
Visualize diffusion
}
\examples{


}
