% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FindPC}
\alias{FindPC}
\title{Determine Number of Principal Components}
\usage{
FindPC(srat, reduction = "pca")
}
\arguments{
\item{srat}{A Seurat object containing the PCA reduction.}

\item{reduction}{Character; the name of the reduction technique to use. Default is "pca".}
}
\value{
Integer; the number of principal components to retain.
}
\description{
Defines the number of principal components (PCs) to retain based on the elbow plot criteria.
The function uses two criteria:
\enumerate{
\item The point where the principal components only contribute 5\% of the standard deviation and the principal components cumulatively contribute 90\% of the standard deviation.
\item The point where the percent change in variation between consecutive PCs is less than 0.1\%.
}
}
\references{
For more details on the elbow plot method, refer to the \href{https://hbctraining.github.io/scRNA-seq/lessons/elbow_plot_metric.html}{Harvard Chan Bioinformatics Core scRNA-seq training}.
}
\keyword{internal}
