% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lmds.R
\name{get_score_profile}
\alias{get_score_profile}
\title{Compute Score Profile}
\usage{
get_score_profile(
  state_0 = NULL,
  state_t = NULL,
  state_inf = NULL,
  P_diag_t = NULL,
  score_ls = c("score0", "delta_correction")
)
}
\arguments{
\item{state_0}{matrix; the initial state matrix.}

\item{state_t}{matrix; the state matrix at time t.}

\item{state_inf}{matrix; the state matrix at equilibrium.}

\item{P_diag_t}{matrix; the diagonal elements of the diffusion operator at time t.}

\item{score_ls}{character vector; a list of scores to compute. Possible values are "score0", "max_score0", "score1", "delta_correction", and "entropy". Default is \code{c("score0", "delta_correction")}.
\itemize{
\item "score0": KL divergence between the initial state (\code{state_0}) and the state at time t (\code{state_t}).
\item "max_score0": KL divergence between the initial state (\code{state_0}) and the equilibrium state (\code{state_inf}).
\item "score1": KL divergence between the state at time t (\code{state_t}) and the equilibrium state (\code{state_inf}).
\item "delta_correction": KL divergence between the initial state (\code{state_0}) and the diagonal elements of the diffusion operator at time t (\code{diag(P^t)}).
\item "entropy": Entropy of the state at time t (\code{state_t}).
}}
}
\value{
A data frame containing the computed scores, with each column representing a score.
}
\description{
Computes various scores for state transitions based on different metrics such as KL divergence and entropy.
}
\keyword{internal}
