% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lmds.R
\name{obtain_lmds}
\alias{obtain_lmds}
\title{Obtain LMD Score (LMDS)}
\usage{
obtain_lmds(score_profile, correction = FALSE)
}
\arguments{
\item{score_profile}{data frame; the score profile containing scores computed at different diffusion times. This is typically the output from \code{fast_calculate_score_profile}, \code{fast_calculate_score_profile_largeData}, or \code{fast_calculate_score_profile_fineTimeScale}.}

\item{correction}{logical; if TRUE, adjusts the score profile by scaling the score profile by entropy and applying delta correction. Default is FALSE.}
}
\value{
A numeric vector containing the LMDS for each row in the score profile (i.e. each gene).
}
\description{
Calculates LMD Score (LMDS) from a given score profile.
}
\keyword{internal}
