% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_knn_graph.R
\name{plot_knn_graph}
\alias{plot_knn_graph}
\title{Plot a KNN graph}
\usage{
plot_knn_graph(affinity_m, label = NULL, layout)
}
\arguments{
\item{affinity_m}{matrix; Symmetry affinity matrix of kNN graph}

\item{label}{string; Node label}

\item{layout}{matrix; 2D coordinate of cells for visualization}
}
\value{
A ggplot object
}
\description{
Plot a KNN graph
}
\examples{


}
