% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{RowwiseNormalize}
\alias{RowwiseNormalize}
\title{Row-wise Normalize a Matrix}
\usage{
RowwiseNormalize(x)
}
\arguments{
\item{x}{matrix; the matrix to be row-wise normalized.}
}
\value{
A numeric matrix with each row normalized so that the sum of the values in each row equals 1.
}
\description{
Normalizes each row of a matrix so that the sum of the values in each row equals 1. Rows with a sum of zero are excluded.
}
