% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lmds.R
\name{show_result_lmd}
\alias{show_result_lmd}
\title{Show LMD Result}
\usage{
show_result_lmd(res.lmd, kneeplot = FALSE)
}
\arguments{
\item{res.lmd}{list; the output of the \code{\link{LMD}}, containing the LMD score profile and LMD scores.}

\item{kneeplot}{logical; if TRUE, a knee plot of the LMD score distribution will be shown.}
}
\value{
A list containing:
\item{gene_table}{data frame; a table with genes, their LMD scores, and their ranks.}
\item{gene_rank}{named numeric vector; the rank of each gene.}
\item{cut_off_gene}{named numeric vector; the top-ranked genes up to the knee point in the LMD score distribution.}
}
\description{
Display the results of LMD.
}
