% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symmetric_KNN_graph.R
\name{Symmetric_KNN_graph}
\alias{Symmetric_KNN_graph}
\title{Construct a symmetric KNN graph}
\usage{
Symmetric_KNN_graph(
  knn = 5,
  feature_space,
  adjust_by_MST = TRUE,
  self_loop = 1
)
}
\arguments{
\item{knn}{integer; knn for constructing the graph}

\item{feature_space}{matrix; cell x coordinate (e.g. 20PCs) matrix}

\item{adjust_by_MST}{boolean; TRUE for connecting disconnected components by Minimum Spanning Trees}

\item{self-loop}{integer; weight for self connect (default 1)}
}
\value{
A list of objects
}
\description{
Construct a symmetric KNN graph
}
\examples{


}
