% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{VisualizeDiffusion}
\alias{VisualizeDiffusion}
\title{Visualize Diffusion}
\usage{
VisualizeDiffusion(
  coord,
  init_state,
  P_ls = NULL,
  W = NULL,
  check_time = NULL,
  gene_name = NULL,
  gene_color = "blue"
)
}
\arguments{
\item{coord}{matrix; 2D coordinates of cells for visualization.}

\item{init_state}{matrix; row-wise normalized expression matrix for the initial state.}

\item{P_ls}{list; list of diffusion operators.}

\item{W}{matrix; symmetric affinity matrix of the graph.}

\item{check_time}{vector; time points selected for visualization, e.g., \code{c(2, 16, 128)}.}

\item{gene_name}{character; the name of the gene for visualization.}

\item{gene_color}{character; the color of expression level for visualization. Default is "blue".}
}
\value{
A ggplot2 object representing a series of plots visualizing the diffusion process.
}
\description{
Creates a series of plots visualizing the diffusion process of a gene's expression over time on a given graph.
}
