% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{VisualizeGeneTSNE}
\alias{VisualizeGeneTSNE}
\title{Visualize Gene t-SNE}
\usage{
VisualizeGeneTSNE(dist, gene_partition = NULL, filtered = TRUE)
}
\arguments{
\item{dist}{matrix; a pairwise distance matrix of genes.}

\item{gene_partition}{factor; a vector indicating the partition of genes into modules. Default is NULL.}

\item{filtered}{logical; if TRUE, filters out some genes for each partition based on the SML method (Parisi et al., 2014). Default is TRUE.}
}
\value{
A ggplot2 object representing the t-SNE visualization of gene partitions.
}
\description{
Creates a t-SNE plot for visualizing gene partitions based on a given distance matrix.
}
\references{
Parisi, F., Strino, F., Nadler, B., & Kluger, Y. (2014). Ranking and combining multiple predictors without labeled data. Proceedings of the National Academy of Sciences, 111(4), 1253-1258. doi:10.1073/pnas.1219097111
}
