% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{VisualizeGraph}
\alias{VisualizeGraph}
\title{Visualize Graph}
\usage{
VisualizeGraph(affinity_m, label = NULL, layout)
}
\arguments{
\item{affinity_m}{matrix; symmetric affinity matrix of the graph.}

\item{label}{vector; node labels (optional). Default is NULL.}

\item{layout}{matrix; a 2D coordinate matrix of cells to visualize the graph.}
}
\value{
A ggplot2 object representing the plotted graph.
}
\description{
Visualizes a graph using a given affinity matrix and node layout.
}
\examples{
affinity_m <- matrix(runif(100), nrow = 10)
layout <- matrix(runif(20), nrow = 10)
label <- sample(letters[1:3], 10, replace = TRUE)
plot <- VisualizeGraph(affinity_m, label, layout)
print(plot)


}
