% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{VisualizeScorePattern}
\alias{VisualizeScorePattern}
\title{Visualize Score Pattern}
\usage{
VisualizeScorePattern(
  score_profile,
  genes = NULL,
  label_class = NULL,
  facet_class = NULL,
  add_point = NULL,
  dyadic = TRUE,
  text = FALSE,
  normalize = FALSE
)
}
\arguments{
\item{score_profile}{matrix; the LMD score profile, output of \code{fast_get_lmds}.}

\item{genes}{vector; selected genes for visualization.}

\item{label_class}{vector; gene class labels for visualizing genes in different colors. Default is NULL.}

\item{facet_class}{vector; gene facet labels for visualizing genes in different panels. Default is NULL.}

\item{add_point}{vector; time points to highlight. Default is NULL.}

\item{dyadic}{logical; if TRUE, represents the axis text in a 2^ format. Default is TRUE.}

\item{text}{logical; if TRUE, adds gene labels to each curve. Default is FALSE.}

\item{normalize}{logical; if TRUE, normalizes the LMD profile by the plateau. Default is FALSE.}
}
\value{
A ggplot2 object representing the score patterns for the selected genes.
}
\description{
Visualizes the LMD score profiles for selected genes over time.
}
