% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddModuleActivityScore.R
\name{AddModuleActivityScore}
\alias{AddModuleActivityScore}
\title{Add module activity scores}
\usage{
AddModuleActivityScore(
  srat,
  gene_partition,
  assay = "RNA",
  do_local_smooth = FALSE,
  knn = 10,
  major_vote = 6,
  nloop = 100,
  module_names = NULL
)
}
\arguments{
\item{srat}{seurat object}

\item{gene_partition}{vector; gene modules}

\item{assay}{string; asssay used for calculated modulescore}

\item{do_local_smooth}{Boolean; TRUE for smooth the module score for each cell by its neighborhoods}

\item{knn}{integer; neighborhoods size for local smoothing}

\item{major_vote}{integer; majority vote number for local smoothing}

\item{nloop}{integer; number of sampling}

\item{module_names}{string; prefix name for each gene module}
}
\value{
A Seurat object
}
\description{
Add module activity scores
}
\examples{


}
